% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-estimates.R
\name{estimates-tab}
\alias{estimates-tab}
\alias{estimatesTab}
\title{Estimates Application Tab}
\usage{
estimatesTab(input, output, session)
}
\arguments{
\item{input}{Shiny input.}

\item{output}{Shiny output.}

\item{session}{Shiny session.}
}
\description{
A tab for the TB model estimate results. This function should not be called
directly. Rather the function is passed to \link{includeTab}.
}
\keyword{internal}
