% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyeglass.R
\docType{package}
\name{eyeglass}
\alias{eyeglass}
\alias{eyeglass-package}
\alias{eyeglass-package}
\title{eyeglass: An Application for Visualizing Model Output}
\description{
More about what it does (maybe more than one line)
}
