% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include-tab.R
\name{include-tab}
\alias{include-tab}
\alias{includeTab}
\title{Include an Application Tab}
\usage{
includeTab(tab, id, ..., session = getDefaultReactiveDomain())
}
\arguments{
\item{tab}{A tab function, much like a module function.}

\item{id}{The id of the tab.}

\item{session}{Mostly ignored.}
}
\description{
Based on shiny's \code{callModule}.
}
\seealso{
\link{estimatesTab}, \link{trendsTab}
}
