% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-preview.R
\name{plot-preview}
\alias{plot-preview}
\alias{plotPreview}
\alias{renderPreview}
\title{Preview A Plot}
\usage{
plotPreview(id)

renderPreview(p)
}
\arguments{
\item{id}{Namespace id.}
}
\description{
A small render of a Shiny plot.
}
\keyword{internal}
