% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-components.R
\name{tab-side-panel}
\alias{tab-side-panel}
\alias{tabSidePanel}
\alias{tabMainPanel}
\title{Eyeglass Application Tab Components}
\usage{
tabSidePanel(id, active = FALSE, comparators = NULL, populations = NULL,
  ages = NULL, years = NULL, outcomes = NULL, interventions = NULL,
  analyses = NULL)

tabMainPanel(id, ..., active = FALSE, dblclick = FALSE, brush = FALSE)
}
\description{
Build out the tabs of the eyeglass application.
}
\keyword{internal}
