% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-trends.R
\name{trends-tab}
\alias{trends-tab}
\alias{trendsTab}
\title{Trends Application Tab}
\usage{
trendsTab(input, output, session)
}
\arguments{
\item{input}{Shiny input.}

\item{output}{Shiny output.}

\item{session}{Shiny session.}
}
\description{
A tab for the TB model trend results. This function should not be called
directly. Rather the function is passed to \link{includeTab}.
}
\keyword{internal}
